# Conflicts:
#	src/main/java/com/example/trs/controller/ActivityController.java
#	src/main/java/com/example/trs/dto/ActivityDTO.java
#	src/main/java/com/example/trs/error/ApiError.java
#	src/main/java/com/example/trs/exceptions/ActivityAlreadyExistsException.java
#	src/main/java/com/example/trs/handler/RestResponseEntityExceptionHandler.java
#	src/main/java/com/example/trs/mapper/ActivityMapper.java
#	src/main/java/com/example/trs/model/Employee.java
#	src/main/java/com/example/trs/repositories/InvoiceJpaRepo.java
#	src/main/java/com/example/trs/service/AbisActivityService.java
#	src/main/java/com/example/trs/service/AbisProjectService.java
#	src/main/java/com/example/trs/service/ActivityService.java
#	src/main/resources/Script.sql

+++ src/main/java/com/example/trs/dto/ProjectDTO.java
@@ -8,7 +8,7 @@ public class ProjectDTO {
  private   String  clientName;
  private   int  clientId;
  private   String name;
- private   int hourlyRate;
+ private   double  hourlyRate;
  private   LocalDate start;
  private   LocalDate  end;
  private    String description;
@@ -48,11 +48,11 @@ public class ProjectDTO {
         this.name = name;
     }
 
-    public int getHourlyRate() {
+    public double getHourlyRate() {
         return hourlyRate;
     }
 
-    public void setHourlyRate(int hourlyRate) {
+    public void setHourlyRate(double hourlyRate) {
         this.hourlyRate = hourlyRate;
     }
